@@ -7,14 +7,14 @@
     "lib": "lib"
   },
   "scripts": {
-    "lint": "eslint **/.js",
+    "lint": "eslint **/*.js",
     "test": "jest --coverage --verbose",
     "test-watch": "jest --watchAll --coverage --verbose"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "eslint": "^5.6.1",
+    "eslint": "^5.16.0",
     "jest": "^23.6.0"
   }
 }
