@@ -1,6 +1,7 @@
 'use strict';
 
-const fileReader = require('./lib/reader.js');
+// const fileReader = require('./lib/reader.js');
+const fileReader = require('./lib/reader-promises.js');
 
 // Obtain and assert input
 let files = process.argv.slice(2);
@@ -11,5 +12,7 @@ if( ! (files instanceof Array && files.length) ) {
 
 fileReader(files, (err,data) => {
   if ( err ) { throw err; }
+  // console.log('here');
   console.log('From Callback:', data);
 });
+
